etnaviv: add acceleration for GetImage()
Provide an accelerated GetImage implementation.  This implmentation
optimises the case where the source pixmap is already on the GPU,
avoiding bouncing it back to the CPU before copying the data.

Instead, we allocate a new pixmap of the desired width and height, use
the GPU to copy to the new pixmap, before falling back to the non-
accelerated methods to copy the data from the new pixmap.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>