etnaviv: cpu/gpu read/write state tracking
Track the state of pixmaps on the CPU and GPU, including whether they
are being read from, or written to by each.  This allows us to permit
concurrent read-only access by the CPU and GPU to a pixmap without
having to map and unmap it from the GPU.

On Dove, this gives an overall 10% performance boost as measured by
gtkperf.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>