etnaviv: delay freeing pixmaps when GPU busy
When we need to free a pixmap which is busy on the GPU, transfer it
to a separate list rather than freeing it.  This avoids an explicit
commit and stall at this point - we can batch up the pixmaps to be
freed, and do that work when we have a pressing reason to commit and
stall.

This batching reduces the number of commit+stall events, and the
resulting interrupt load - the number of interrupts received during
a gtkperf run is roughly halved.

Limit the maximum time that unused pixmaps can be on the busy_free_list
to 500ms, before we invoke a commit+stall event.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>