vivante: dri2: fix flips
The flip code suffers from two issues.  Firstly, the flip is scheduled
immediately even if the requested msc is in the future.  We should only
schedule the flip immediately if we have reached the msc or the msc is
in the past.

Secondly, when we schedule the flip immediately, we should update the
target msc, otherwise it remains at its present value indefinitely.

Lastly, we micro-optimise the target_msc update in the "future" case -
there is no need to update the value if the value we're going to write
back is the same as its present value.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>