etnaviv: Correct etnaviv_de_op flush numbers
The math behind when etnaviv_de_op should flush is incorrect, and can
lead to assert failures when issuing a large number of draws.  This
was particularly apparent in Icedove, which could cause the X server
to abort due to an assert failure.

The correct math should be 3 + 3 * nBox, not 2 + 2 * nBox.

Signed-off-by: Sean Cross <xobs@kosagi.com>